[cleanup] Decrease complexity by segmenting table refresh class.
